# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tech-Admin-Api",
        "contact": {
            "name": "Tech-Admin",
            "url": "  http://tech-admin.com",
            "email": "  @gmail.com"
        },
        "version": "0.1"
    },
    "paths": {
...
        "/api/acesso/{id}": {
            "get": {
                "tags": [
                    "Acesso"
                ],
                "summary": "Mostra um acesso",
                "description": "Mostra um acesso",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do acesso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna o acesso"
                    },
                    "default": {
                        "description": "Caso não encontre o acesso"
                    }
                }
            },
            "put": {
                "tags": [
                    "Acesso"
                ],
                "summary": "Atualiza um acesso",
                "description": "Atualiza um acesso",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do acesso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "description": "Nome do acesso",
                                        "type": "string"
                                    },
                                    "descricao": {
                                        "description": "Descrição do acesso",
                                        "type": "string"
                                    },
                                    "status": {
                                        "description": "Status do acesso",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retorna o acesso atualizado"
                    },
                    "default": {
                        "description": "Caso não encontre o acesso"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Acesso"
                ],
                "summary": "Deleta um acesso",
                "description": "Deleta um acesso",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do acesso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna o acesso deletado"
                    },
                    "default": {
                        "description": "Caso não encontre o acesso"
                    }
                }
            }
        },
...
        "/api/endereco": {
            "get": {
                "tags": [
                    "Endereco"
                ],
                "summary": "Lista todos os endereços",
                "description": "Retorna todos os endereços",
                "operationId": "indexEndereco",
                "responses": {
                    "200": {
                        "description": "Retorna todos os endereços"
                    },
                    "default": {
                        "description": "Caso não encontre o endereço"
                    }
                }
            },
            "post": {
                "tags": [
                    "Endereco"
                ],
                "summary": "Cria um novo endereço",
                "description": "Cria um novo endereço",
                "operationId": "storeEndereco",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "logradouro",
                                    "numero",
                                    "bairro",
                                    "cidade",
                                    "estado",
                                    "cep"
                                ],
                                "properties": {
                                    "logradouro": {
                                        "description": "Logradouro do endereço",
                                        "type": "string"
                                    },
                                    "numero": {
                                        "description": "Número do endereço",
                                        "type": "string"
                                    },
                                    "bairro": {
                                        "description": "Bairro do endereço",
                                        "type": "string"
                                    },
                                    "cidade": {
                                        "description": "Cidade do endereço",
                                        "type": "string"
                                    },
                                    "estado": {
                                        "description": "Estado do endereço",
                                        "type": "string"
                                    },
                                    "cep": {
                                        "description": "CEP do endereço",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retorna o endereço criado"
                    },
                    "default": {
                        "description": "Caso não encontre o endereço"
                    }
                }
            }
        },
        "/api/endereco/{id}": {
            "get": {
                "tags": [
                    "Endereco"
                ],
                "summary": "Mostra um endereço",
                "description": "Mostra um endereço",
                "operationId": "showEndereco",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do endereço",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna o endereço"
                    },
                    "default": {
                        "description": "Caso não encontre o endereço"
                    }
                }
            },
            "put": {
                "tags": [
                    "Endereco"
                ],
                "summary": "Atualiza um endereço",
                "description": "Atualiza um endereço",
                "operationId": "updateEndereco",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do endereço",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "logradouro",
                                    "numero",
                                    "bairro",
                                    "cidade",
                                    "estado",
                                    "cep"
                                ],
                                "properties": {
                                    "logradouro": {
                                        "description": "Logradouro do endereço",
                                        "type": "string"
                                    },
                                    "numero": {
                                        "description": "Número do endereço",
                                        "type": "string"
                                    },
                                    "bairro": {
                                        "description": "Bairro do endereço",
                                        "type": "string"
                                    },
                                    "cidade": {
                                        "description": "Cidade do endereço",
                                        "type": "string"
                                    },
                                    "estado": {
                                        "description": "Estado do endereço",
                                        "type": "string"
                                    },
                                    "cep": {
                                        "description": "CEP do endereço",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retorna o endereço atualizado"
                    },
                    "default": {
                        "description": "Caso não encontre o endereço"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Endereco"
                ],
                "summary": "Deleta um endereço",
                "description": "Deleta um endereço",
                "operationId": "deleteEndereco",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do endereço",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna o endereço deletado"
                    },
                    "default": {
                        "description": "Caso não encontre o endereço"
                    }
                }
            }
        },
...
    }
}